Move TranslationSourceType enum to API; remove the class
