+++ api/zanata-common-api/src/main/java/org/zanata/Common.gwt.xml
@@ -6,5 +6,6 @@
   <source path="rest/dto">
     <include name="stats/*" />
     <include name="Link*" />
+    <include name="TranslationSourceType*" />
   </source>
 </module>
