+++ server/gwt-shared/src/main/java/org/zanata/webtrans/shared/model/TransUnit.java
@@ -8,6 +8,7 @@ import java.util.List;
 import com.google.common.base.MoreObjects;
 import org.zanata.common.ContentState;
 import org.zanata.common.LocaleId;
+import org.zanata.rest.dto.TranslationSourceType;
 
 import com.google.common.base.Preconditions;
 import com.google.common.base.Strings;
