+++ server/services/src/main/java/org/zanata/rest/editor/service/StatisticsService.java
@@ -32,10 +32,9 @@ import javax.inject.Named;
 import org.zanata.common.LocaleId;
 import org.zanata.dao.DocumentDAO;
 import org.zanata.model.HDocument;
-import org.zanata.model.type.TranslationSourceType;
+import org.zanata.rest.dto.TranslationSourceType;
 import org.zanata.rest.RestUtil;
 import org.zanata.rest.dto.stats.ContainerTranslationStatistics;
-import org.zanata.rest.dto.stats.TranslationStatistics;
 import org.zanata.rest.dto.stats.TranslationStatistics.StatUnit;
 import org.zanata.rest.editor.dto.EditorTranslationStatistics;
 import org.zanata.rest.editor.service.resource.StatisticResource;
