+++ server/services/src/main/java/org/zanata/service/impl/MachineTranslationServiceImpl.java
@@ -58,7 +58,7 @@ import org.zanata.model.HLocale;
 import org.zanata.model.HProjectIteration;
 import org.zanata.model.HTextFlow;
 import org.zanata.model.HTextFlowTarget;
-import org.zanata.model.type.TranslationSourceType;
+import org.zanata.rest.dto.TranslationSourceType;
 import org.zanata.rest.dto.MachineTranslationPrefill;
 import org.zanata.service.LocaleService;
 import org.zanata.service.MachineTranslationService;
