+++ server/services/src/test/java/org/zanata/dao/TextFlowTargetDAOTest.java
@@ -28,7 +28,7 @@ import org.zanata.ZanataDbunitJpaTest;
 import org.zanata.model.HTextFlowTarget;
 import org.zanata.model.HTextFlowTargetHistory;
 import org.zanata.model.type.EntityType;
-import org.zanata.model.type.TranslationSourceType;
+import org.zanata.rest.dto.TranslationSourceType;
 
 import static org.assertj.core.api.Assertions.assertThat;
