+++ server/services/src/test/java/org/zanata/search/FilterConstraintToQueryTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
 import static org.mockito.Mockito.when;
-import static org.zanata.model.type.TranslationSourceType.MACHINE_TRANS;
+import static org.zanata.rest.dto.TranslationSourceType.MACHINE_TRANS;
 import static org.zanata.search.FilterConstraintToQuery.Parameters.*;
 import java.util.List;
 import org.hibernate.Query;
@@ -13,7 +13,6 @@ import org.junit.Test;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 import org.zanata.model.HLocale;
-import org.zanata.webtrans.shared.model.ContentStateGroup;
 import org.zanata.webtrans.shared.model.DocumentId;
 import org.zanata.webtrans.shared.search.FilterConstraints;
 import com.google.common.collect.Lists;
