+++ server/services/src/test/java/org/zanata/service/impl/CopyVersionServiceImplTest.java
@@ -64,8 +64,7 @@ import org.zanata.model.HTextFlowTarget;
 import org.zanata.model.HTextFlowTargetHistory;
 import org.zanata.model.HTextFlowTargetReviewComment;
 import org.zanata.model.po.HPoTargetHeader;
-import org.zanata.model.type.TranslationSourceType;
-import org.zanata.seam.security.CurrentUserImpl;
+import org.zanata.rest.dto.TranslationSourceType;
 import org.zanata.security.ZanataCredentials;
 import org.zanata.security.ZanataIdentity;
 import com.google.common.collect.Lists;
