+++ server/zanata-model/src/main/java/org/zanata/model/type/TranslationSourceTypeType.java
@@ -26,6 +26,7 @@ import org.hibernate.dialect.Dialect;
 import org.hibernate.type.AbstractSingleColumnStandardBasicType;
 import org.hibernate.type.DiscriminatorType;
 import org.hibernate.type.StringType;
+import org.zanata.rest.dto.TranslationSourceType;
 
 
 public class TranslationSourceTypeType extends AbstractSingleColumnStandardBasicType<TranslationSourceType>
