+++ server/zanata-model/src/main/java/org/zanata/model/type/TranslationSourceTypeTypeDescriptor.java
@@ -23,6 +23,7 @@ package org.zanata.model.type;
 
 import org.hibernate.type.descriptor.WrapperOptions;
 import org.hibernate.type.descriptor.java.AbstractTypeDescriptor;
+import org.zanata.rest.dto.TranslationSourceType;
 
 public class TranslationSourceTypeTypeDescriptor extends
         AbstractTypeDescriptor<TranslationSourceType> {
